Code review fixes
Custom editor moved to a different file along with a new asmdef, and other fixes requested.
