+++ Explorer/Assets/DCL/Roads/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f6c9e3a0b311fb42b708efb23fc572e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
