+++ Explorer/Assets/DCL/Roads/Editor/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0aaa4a1bda391674ebbcefc8bbfdfd04
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
