+++ Explorer/Assets/DCL/Roads/Settings/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e615bcead634acf43a9d86c2517d5b66
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
