+++ Explorer/Assets/DCL/Roads/Settings/Editor/DCL.Roads.Settings.Editor.asmdef
@@ -1,0 +1,19 @@
+{
+    "name": "DCL.Roads.Settings.Editor",
+    "rootNamespace": "",
+    "references": [
+        "GUID:54cb1906aeb4e4264bc1d2aa3818a43f",
+        "GUID:46c2e553ecab9ff4784aee64075136c9"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
