+++ Explorer/Assets/DCL/Roads/Settings/Editor/DCL.Roads.Settings.Editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 05c5e091fc59a804d8f972e6f29b0c85
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
